worker/requirements.txt
# Core ML
torch>=2.3.0
diffusers>=0.31.0
transformers>=4.44.0
accelerate>=0.33.0
safetensors>=0.4.0
sentencepiece>=0.2.0

# API
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Redis
redis>=5.0.0

# Database
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# R2 / S3
boto3>=1.34.0

# Image / Video
Pillow>=10.0.0
numpy>=1.26.0

# HuggingFace model downloads
huggingface-hub>=0.25.0

# Env
python-dotenv>=1.0.0
